React Star Wars Carousel
This is a refactored version of the React Sprint with a Netflix-esque carousel in React hooks. React Star Wars Carousel is currently two separately deployed apps using create-react-app which will be merged in to one final application. It calls data from the star wars API and maps character qualities and App2 is a star wars movie carousel designed like a netflix movie list. This is a creative commons project using images from star wars and is not intended for anything other than the challenge, not profit.